backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-google-genai
langchain-community
tavily-python
python-dotenv
pydantic>=2.7.4
httpx
